requirements.txt
# REDCap Data Quality Intelligence Agent
# Dependencies

# HTTP requests for REDCap API
requests>=2.31.0

# Data manipulation
pandas>=2.0.0

# Environment variables
# Environment variables
python-dotenv>=1.0.0


# Data validation models
pydantic>=2.0.0

# Rich CLI interface
rich>=13.0.0

# Date/time handling
python-dateutil>=2.8.0

# JSON schema validation
jsonschema>=4.0.0

# PDF Generation
reportlab>=4.0.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Charts for PDF
matplotlib>=3.8.0

# LangChain for AI Agent
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0

# Authentication
supabase>=2.0.0

# Testing
pytest>=7.0.0

# Production Server (Windows)
waitress>=2.1.0

# Production Server (Linux/Cloud)
gunicorn>=20.1.0
